Generalized controller functions to avoid code repetition

+++ sql/CreateTableProducts.sql
@@ -3,7 +3,7 @@ DROP TABLE IF EXISTS NodeJSTraining.Products;
 CREATE TABLE NodeJSTraining.Products (
 	ID INT NOT NULL AUTO_INCREMENT,
 	Name VARCHAR(255) NOT NULL,
-	Price int NOT NULL,
+	Price DECIMAL(18,2) NOT NULL,
 	Quantity int	NOT NULL,
 	PRIMARY KEY (ID)
 );
